requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
coveralls==1.8.2
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
djangorestframework==3.10.3
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.8
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.20
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
Pillow==6.1.0
pluggy==0.12.0
psycopg2==2.7.7
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.3
virtualenv==16.7.5
whitenoise==4.1.3
zipp==0.6.0
